Rc2.1

+++ src/interfaces/python/examples/python_visitor.py
@@ -1,7 +1,7 @@
 import opengm
 import numpy
 shape=[40,40]
-unaries=numpy.random.rand(shape[0], shape[1],2).astype(numpy.float32)
+unaries=numpy.random.rand(shape[0], shape[1],2)
 potts=opengm.PottsFunction([2,2],0.0,0.4)
 gm=opengm.grid2d2Order(unaries=unaries,regularizer=potts)
